Remove monogen package and move it to the cmd/monogen module
