+++ cmd/monogen/testapp/app/genesis_account.go
