+++ cmd/monogen/testapp/buf.work.yaml
