+++ cmd/monogen/testapp/docs/static/openapi.yml
