+++ cmd/monogen/testapp/docs/template/index.tpl
