+++ cmd/monogen/testapp/proto/buf.gen.gogo.yaml
