+++ cmd/monogen/testapp/proto/buf.gen.pulsar.yaml
