+++ cmd/monogen/testapp/proto/buf.gen.sta.yaml
