+++ cmd/monogen/testapp/proto/buf.gen.swagger.yaml
