+++ cmd/monogen/testapp/proto/buf.gen.ts.yaml
