+++ cmd/monogen/testapp/proto/buf.yaml
