+++ cmd/monogen/testapp/setup-helper.sh
