+++ cmd/monogen/testapp/wallet/bech32.js
