+++ cmd/monogen/testapp/wallet/index.html
