+++ cmd/monogen/testapp/wallet/keplr-integration.js
