+++ cmd/monogen/testapp/wallet/metamask-integration.js
